+++ __init__.py
@@ -15,7 +15,7 @@ Installation:
 3. Restart ComfyUI
 
 Author: Sumit Chatterjee
-Version: 1.0.0
+Version: 1.0.1
 License: MIT
 """